requirements.txt
langchain
langchain-community
langchain-openai
chromadb
openai
tiktoken
python-dotenv
pymupdf
